+++ BabelBrain/BabelDatasetPreps.py
@@ -465,7 +465,7 @@ def GetSkullMaskFromSimbNIBSSTL(SimbNIBSDir='4007/4007_keep/m2m_4007_keep/',
             #while charm is much more powerful to segment skull regions, we need to calculate the meshes ourselves
             charminput = inputfilenames['SimbNIBSinput']
             charm = S1_file_manager.load_file(charminput)
-            charmdata=np.ascontiguousarray(charm.get_fdata())[:,:,:,0]
+            charmdata=np.ascontiguousarray(charm.get_fdata(dtype=np.float32))[:,:,:,0]
             AllTissueRegion=charmdata>0 #this mimics what the old headreco does for skin
             
             tissues = S1_file_manager.load_file(charminput)
@@ -793,7 +793,7 @@ def GetSkullMaskFromSimbNIBSSTL(SimbNIBSDir='4007/4007_keep/m2m_4007_keep/',
                     rCT = S1_file_manager.load_file(outputfilenames['T1WinCT'])
                     S1_file_manager.save_file(file_data=fct,filename=outputfilenames['ReuseMask'],precursor_files=outputfilenames['T1WinCT'])
 
-            rCTdata=rCT.get_fdata()
+            rCTdata=rCT.get_fdata(dtype=np.float32)
         else:
             if CTType in [2,3]:
                 print('Processing ZTE/PETRA to pCT')
@@ -827,7 +827,7 @@ def GetSkullMaskFromSimbNIBSSTL(SimbNIBSDir='4007/4007_keep/m2m_4007_keep/',
                                                    ResampleFilter,
                                                    ResampleFilterCOMPUTING_BACKEND)
                 gc.collect()
-            rCTdata=rCT.get_fdata()
+            rCTdata=rCT.get_fdata(dtype=np.float32)
             hist = np.histogram(rCTdata[rCTdata>TypeThresold],bins=15)
             print('*'*40)
             if CTType in [1,2,3]:
@@ -1146,7 +1146,7 @@ def GetSkullMaskFromSimbNIBSSTL(SimbNIBSDir='4007/4007_keep/m2m_4007_keep/',
                 emptyNifti.to_filename(ename)
                 outname = os.path.join(tmpdirname,'out.nii.gz')
                 RunMeshConv(ename,mshfile,outname,SimbNINBSRoot=SimbNINBSRoot)
-                upScaleMask=nibabel.load(outname).get_fdata(dtype=np.int8)
+                upScaleMask=nibabel.load(outname).get_fdata().astype(np.int8)
                 gc.collect()
         
         FinalMask2=FinalMask.copy()
@@ -1166,9 +1166,9 @@ def GetSkullMaskFromSimbNIBSSTL(SimbNIBSDir='4007/4007_keep/m2m_4007_keep/',
     
     with CodeTimer("CTS:L3:S1: resampling T1 to mask",unit='s'):
         if ResampleFilter is None:
-            T1Conformal=processing.resample_from_to(T1Conformal,mask_nifti2,mode='constant',order=0,cval=T1Conformal.get_fdata().min())
+            T1Conformal=processing.resample_from_to(T1Conformal,mask_nifti2,mode='constant',order=0,cval=T1Conformal.get_fdata(dtype=np.float32).min())
         else:
-            T1Conformal=ResampleFilter(T1Conformal,mask_nifti2,mode='constant',order=0,cval=T1Conformal.get_fdata().min(),GPUBackend=ResampleFilterCOMPUTING_BACKEND)
+            T1Conformal=ResampleFilter(T1Conformal,mask_nifti2,mode='constant',order=0,cval=T1Conformal.get_fdata(dtype=np.float32).min(),GPUBackend=ResampleFilterCOMPUTING_BACKEND)
         T1W_resampled_fname=os.path.dirname(T1Conformal_nii)+os.sep+prefix+'T1W_Resampled.nii.gz'
         S1_file_manager.save_file(file_data=T1Conformal,filename=T1W_resampled_fname)
